+++ examples/users-managment/src/main.rs
@@ -1,0 +1,19 @@
+pub mod shared;
+pub mod users;
+
+use dotenv::dotenv;
+use sword::prelude::*;
+
+use crate::{shared::SharedModule, users::UsersModule};
+
+#[sword::main]
+async fn main() {
+    dotenv().ok();
+
+    let app = Application::builder()
+        .with_module::<SharedModule>()
+        .with_module::<UsersModule>()
+        .build();
+
+    app.run().await;
+}