+++ sword-tests/src/cors/mod.rs
@@ -1,0 +1,6 @@
+use sword::core::Application;
+
+#[tokio::test]
+async fn test_cors() {
+    Application::builder().build();
+}